requirements.txt
bcrypt==4.0.1
fastapi
passlib[bcrypt]
psycopg2-binary
python-dotenv
python-jose[cryptography]
python-multipart
requests
sqlalchemy
uvicorn
uvicorn[standard]
redis
psutil
google-generativeai